refactor: adds Spinner and changes the order of paramters in changeName handler also clean up #146
